move unique key terms cache to utils
