+++ comb_spec_searcher/strategies/rule.py
@@ -37,11 +37,11 @@ from comb_spec_searcher.typing import (
     SubObjects,
     SubTerms,
     Terms,
-    TermsCache,
 )
 
 from ..combinatorial_class import CombinatorialClassType, CombinatorialObjectType
 from ..exception import SanityCheckFailure, StrategyDoesNotApply
+from ..utils import TermsCache
 from .constructor import Complement, Constructor, DisjointUnion
 
 if TYPE_CHECKING:
